quick fixes to build the latest version with MSVC
git-svn-id: file://localhost/home/tolik/svnrepo/trunk@1617 c5418bc3-7546-4aae-8bad-fa6f2f3e53e6

+++ opencv/src/cxcore/cxalloc.cpp
@@ -482,7 +482,7 @@ void fastFree( void* ptr )
     {
     ThreadData* tls = ThreadData::get();
     Node* node = (Node*)ptr;
-    Block* block = (Block*)((size_t)ptr & -MEM_BLOCK_SIZE);
+    Block* block = (Block*)((size_t)ptr & (size_t)-MEM_BLOCK_SIZE);
     assert( block->signature == MEM_BLOCK_SIGNATURE );
 
     if( block->threadData == tls )
